Moved the rc file to the app name in a vain attempt to get the mythical class of XMLGUI to maby work.
Removed the listview junk to view properties since its to bugy and not
worth it at this point.

Cleaned up the source so its ready to move to KDESDK.

Rich, if you are happy with this I think I have not seen a bug in it in
a few weeks.  I think its ready for prime time.

CCMAIL: rich@kde.org

svn path=/trunk/kdesdk/kuiviewer/; revision=256374