Move common parts to the top-level CMakeLists.txt
Summary:
Besides de-duplication being a nice thing this also solves
translations getting installed into an undesired location. It
happens because at the time "ki18n_install(po)" (added on tarball
generation in order to install translations) is processed,
KDEInstallDirs is not yet included and LOCALE_INSTALL_DIR is not
defined, which leads to ki18n_install setting it to a default value.

Reviewers: #kde_applications, ltoscano

Differential Revision: https://phabricator.kde.org/D5514