Correctly locate the KPart library now that it is in a subdirectory
As a result of commit 476348de.  To avoid having to also hardcode the
path in the part loader, the part service file is found by its desktop
file name which will automatically take account of the subdirectory.

Fix the error detection so that if the part cannot be found (which is an
installation error) then after the error message there is no crash; the
application simply exits as intended.